Generate memory id from when users enter the llm search page to when users submit an input"
