+++ src/features/LlmSearch/SearchPane.jsx
@@ -3,24 +3,42 @@ import React, { useState } from "react";
 import Box from "@mui/joy/Box";
 import Sheet from "@mui/joy/Sheet";
 import Stack from "@mui/joy/Stack";
+import Typography from "@mui/joy/Typography";
 
 import MessageBubble from "./MessageBubble";
 import SearchInput from "./SearchInput";
 
 import { NO_HEADER_BODY_HEIGHT } from "../../configs/VarConfigs";
-import { fetchLlmSearchResult } from "../../utils/DataRetrieval";
-import { Typography } from "@mui/material";
+import {
+  fetchLlmSearchResult,
+  fetchLlmSearchMemoryId,
+} from "../../utils/DataRetrieval";
+import { SampleChats } from "./SampleChats";
 
-const TEST_MODE = import.meta.env.VITE_TEST_MODE;
+const DO_NOT_USE_LLM_ENDPOINT = import.meta.env.VITE_DO_NOT_USE_LLM_ENDPOINT;
+
+async function fetchMemoryIdForNewConversation() {
+  const llmMemory =
+    DO_NOT_USE_LLM_ENDPOINT === "true" ? {} : await fetchLlmSearchMemoryId();
+
+  if (llmMemory === "ERROR") {
+    setError(true);
+    return;
+  }
+
+  return llmMemory.memoryId;
+}
 
 // Get llm search result and update chatMessages array
 async function getLlmSearchResult(
   input,
   memoryId,
+  setMemoryId,
   chatMessages,
   setChatMessages,
   setWaitingForResponse
 ) {
+  let result;
   const newId = chatMessages.length + 1;
   const newIdString = newId.toString();
 
@@ -33,8 +51,13 @@ async function getLlmSearchResult(
   setWaitingForResponse(true);
   setChatMessages(currentChatMessage);
 
-  const result = await fetchLlmSearchResult(input, memoryId);
-  TEST_MODE && console.log("Question", input, memoryId, "Response", result);
+  if (!memoryId) {
+    const newMemoryId = await fetchMemoryIdForNewConversation();
+    setMemoryId(newMemoryId);
+    result = await fetchLlmSearchResult(input, newMemoryId);
+  } else {
+    result = await fetchLlmSearchResult(input, memoryId);
+  }
 
   if (!result) {
     alert("Error getting response from I-GUIDE AI.");
@@ -57,9 +80,20 @@ async function getLlmSearchResult(
 }
 
 export default function SearchPane(props) {
-  const startingChat = props.startingChat;
-  const memoryId = props.memoryId;
+  // Greetings from I-GUIDE
+  const starterChat = [
+    [
+      {
+        message_id: "0",
+        answer: "Hi! What can I help you today :)",
+        sender: "I-GUIDE AI",
+      },
+    ],
+  ];
+  const startingChat =
+    DO_NOT_USE_LLM_ENDPOINT === "true" ? SampleChats[0] : starterChat[0];
 
+  const [memoryId, setMemoryId] = useState("");
   const [chatMessages, setChatMessages] = useState(startingChat);
   const [searchInputValue, setSearchInputValue] = useState("");
   const [waitingForResponse, setWaitingForResponse] = useState(false);
@@ -122,6 +156,7 @@ export default function SearchPane(props) {
           getLlmSearchResult(
             searchInputValue,
             memoryId,
+            setMemoryId,
             chatMessages,
             setChatMessages,
             setWaitingForResponse
